% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioacc_metric.R
\name{bioacc_metric}
\alias{bioacc_metric}
\alias{bioacc_metric.fitTK}
\title{Biaccumulation metrics}
\usage{
bioacc_metric(fit, ...)

\method{bioacc_metric}{fitTK}(fit, object)
}
\description{
Biaccumulation metrics
}
