% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioacc-data.R
\docType{data}
\name{Chiro_Creuzot}
\alias{Chiro_Creuzot}
\title{Data on Chironomus with several exposure routes.}
\format{
A dataframe with 24 observations on the following four variables:
\describe{
\item{\code{time}}{A vector of class \code{numeric} with the time points in
days.}
\item{\code{expw}}{A vector of class \code{numeric} with the exposure
in water.}
\item{\code{expw}}{A vector of class \code{numeric} with the exposure
in pore water.}
\item{\code{replicate}}{A vector of class \code{integer} for replicate
identification.}
\item{\code{conc}}{A vector of class \code{numeric} with concentration
in organism.}
\item{\code{concm1}}{A vector of class \code{numeric} with metabolite
concentration in organism.}
\item{\code{concm2}}{A vector of class \code{numeric} with metabolite
concentration in organism.}
}
}
\usage{
data(Chiro_Creuzot)
}
\description{
Data on Chironomus with several exposure routes.
}
\keyword{dataset}
