% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioacc-data.R
\docType{data}
\name{Chironomus_benzoapyrene}
\alias{Chironomus_benzoapyrene}
\title{Data on Chironomus exposed to benzoapyrene}
\usage{
data(Chironomus_benzoapyrene)
}
\description{
Data on Chironomus exposed to benzoapyrene
}
\keyword{dataset}
