% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexTK_stan.r
\name{complexTK_stan}
\alias{complexTK_stan}
\title{Bayesian inference of simple model with Stan}
\usage{
complexTK_stan(stanTKdata, ...)
}
\arguments{
\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Bayesian inference of simple model with Stan
}
