% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_fit.R
\name{corrMatrix}
\alias{corrMatrix}
\title{Correlations between parameters: colored matrix}
\usage{
corrMatrix(fit)
}
\arguments{
\item{fit}{An object of class \code{fitTK}}
}
\value{
A heatmap of class \code{ggplot}.
}
\description{
Correlations between parameters: colored matrix
}
