% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_fit.R
\name{corrPlot}
\alias{corrPlot}
\title{Correlations between parameters: pairs plot}
\usage{
corrPlot(fit, plots = c("all", "deterministic", "stochastic"))
}
\arguments{
\item{fit}{An object of class \code{fitTK}}

\item{plots}{A string selecting the parameters. Defaults is \code{"all"} and select all parameters.
Deterministc parameters can be selected by setting \code{"deterministic"} and
stochastic parameter with \code{"stochastic"}}
}
\value{
A pairsplot of class \code{ggmatrix} containing planes of parameter pairs (lower triangle), marginal posterior distribution of each parameter (diagonal) and Pearson correlation coefficients (upper triangle)
}
\description{
Correlations between parameters: pairs plot
}
