% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PP.R
\name{df_PriorPost}
\alias{df_PriorPost}
\alias{df_PriorPost.fitTK}
\title{Data frame of Posterior over Prior}
\usage{
df_PriorPost(fit, ...)

\method{df_PriorPost}{fitTK}(fit, select = "all", ...)
}
\arguments{
\item{fit}{An object of class \code{fitTK} returned by the function \code{fitTK()}.}

\item{\dots}{Additional arguments}

\item{select}{A string selecting the parameters. Defaults is \code{"all"} and
select all parameters.Deterministc parameters can be selected by setting
\code{"deterministic"} and stochastic parameter with \code{"stochastic"}}
}
\value{
An object of class \code{data.frame}
}
\description{
Data frame of Posterior over Prior

Data frame of Posterior over Prior
}
