% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_internals.R
\name{.fonte}
\alias{.fonte}
\title{A simple implementation of to pivot_longer of tidyr}
\usage{
.fonte(df, names_to, values_to)
}
\arguments{
\item{df}{A data frame to pivot.}

\item{names_to}{A string specifying the name of the column to create from
the data stored in the column names of \code{df}.}

\item{values_to}{A string specifying the name of the column to create from
the data stored in cell values.}
}
\value{
The data frame with a "lengthens" shape: more rows, less columns
}
\description{
A simple implementation of to pivot_longer of tidyr
}
