% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_internals.R
\name{.index_col_exposure}
\alias{.index_col_exposure}
\title{Return column matching "expw", "exps", "expf", "exppw" of a \code{data.frame}}
\usage{
.index_col_exposure(data_frame)
}
\arguments{
\item{data_frame}{a dataframe}
}
\value{
A vector of numeric
}
\description{
Return column matching "expw", "exps", "expf", "exppw" of a \code{data.frame}
}
