% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_internals.R
\name{.index_col_metabolite}
\alias{.index_col_metabolite}
\title{Return column matching "concX" of a \code{data.frame} where X is metabolite}
\usage{
.index_col_metabolite(data_frame)
}
\arguments{
\item{data_frame}{a dataframe}
}
\value{
A vector of numeric
}
\description{
Return column matching "concX" of a \code{data.frame} where X is metabolite
}
