% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_internals.R
\name{.is_equal_rmInf}
\alias{.is_equal_rmInf}
\title{Check if two vectors \code{x} and \code{y} are equal after remove \code{Inf}}
\usage{
.is_equal_rmInf(x, y)
}
\arguments{
\item{x}{A vector}

\item{y}{A vector}
}
\value{
A logical value
}
\description{
Check if two vectors \code{x} and \code{y} are equal after remove \code{Inf}
}
