% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations.R
\name{equations}
\alias{equations}
\title{Equations of the mathematical model used for the fit}
\usage{
equations(fit, object)
}
\arguments{
\item{fit}{An object of class \code{fitTK}}

\item{object}{The data.frame used as the base as the fit object}
}
\value{
A vector of strings each containing an equation
}
\description{
Equations of the mathematical model used for the fit
}
