% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioacc_metric.R
\name{exposure_names}
\alias{exposure_names}
\title{Retrieve exposure routes names from object}
\usage{
exposure_names(object)
}
\arguments{
\item{object}{a data frame.}
}
\value{
A vector of string
}
\description{
Retrieve exposure routes names from object
}
