% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioacc-data.R
\docType{data}
\name{Exposure_Sialis_lutaria}
\alias{Exposure_Sialis_lutaria}
\title{Data on Sialis lutaria exposure time series}
\usage{
data(Exposure_Sialis_lutaria)
}
\description{
Data on Sialis lutaria exposure time series
}
\keyword{dataset}
