% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTK.R
\name{fitTK2}
\alias{fitTK2}
\title{Bayesian inference of TK model with Stan}
\usage{
fitTK2(stanTKdata, ...)
}
\arguments{
\item{stanTKdata}{List of Data require for computing}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{fitTK} returned by \code{rstan::sampling}
}
\description{
Bayesian inference of TK model with Stan
}
