% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioacc-data.R
\docType{data}
\name{Gammarus_azoxistrobine_1d_Rosch2017}
\alias{Gammarus_azoxistrobine_1d_Rosch2017}
\title{Data on Gammarus exposed to azoxistrobine}
\usage{
data(Gammarus_azoxistrobine_1d_Rosch2017)
}
\description{
Data on Gammarus exposed to azoxistrobine
}
\keyword{dataset}
