% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioacc-data.R
\docType{data}
\name{Internal_Sialis_lutaria}
\alias{Internal_Sialis_lutaria}
\title{Data on Sialis lutaria internal time series}
\usage{
data(Internal_Sialis_lutaria)
}
\description{
Data on Sialis lutaria internal time series
}
\keyword{dataset}
