% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioacc-data.R
\docType{data}
\name{Male_Gammarus_seanine_growth}
\alias{Male_Gammarus_seanine_growth}
\title{Male \code{Gammarus pulex} exposed to seanine spiked water. A single exposure
concentration was tested. The duration of the accumulation phase is 1.417
days. Three metabolites were quantified. The growth of organism was included.}
\format{
A dataframe with 22 observations on the following four variables:
\describe{
\item{\code{time}}{A vector of class \code{numeric} with the time points in
days.}
\item{\code{expw}}{A vector of class \code{numeric} with seanine
exposure in water in \eqn{\mu g.m L^{-1}}.}
\item{\code{replicate}}{A vector of class \code{integer} for replicate
identification.}
\item{\code{conc}}{A vector of class \code{numeric} with concentration
in organism.}
\item{\code{concm1}}{A vector of class \code{numeric} with metabolite
concentration in organism.}
\item{\code{concm2}}{A vector of class \code{numeric} with metabolite
concentration in organism.}
\item{\code{concm3}}{A vector of class \code{numeric} with metabolite
concentration in organism.}
\item{\code{growth}}{A vector of class \code{numeric} with growth of the
organism.}
}
}
\usage{
data(Male_Gammarus_seanine_growth)
}
\description{
Male \code{Gammarus pulex} exposed to seanine spiked water. A single exposure
concentration was tested. The duration of the accumulation phase is 1.417
days. Three metabolites were quantified. The growth of organism was included.
}
\references{
Ashauer, R. et al. (2012). Significance of xenobiotic metabolism
for bioaccumulation kinetics of organic chemicals in Gammarus pulex.
Environmental Science Technology, 46: 3498-3508.
}
\keyword{dataset}
