% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioacc-data.R
\docType{data}
\name{Male_Gammarus_Single}
\alias{Male_Gammarus_Single}
\title{Bio-accumulation data set for \emph{Gammarus fossarum} exposed to Hg spiked water.}
\format{
A dataframe with 23 observations on the following four variables:
\describe{
\item{\code{time}}{A vector of class \code{numeric} with the time points in
days.}
\item{\code{expw}}{A vector of class \code{numeric} with Hg
exposure in water in \eqn{\mu g.m L^{-1}}.}
\item{\code{replicate}}{A vector of class \code{integer} for replicate
identification.}
\item{\code{conc}}{A vector of class \code{numeric} with Hg concentration
in organism in \eqn{\mu g.m L^{-1}}.}
}
}
\usage{
data(Male_Gammarus_Single)
}
\description{
Male \emph{Gammarus fossarum} exposed to Hg spiked water.
A single exposure concentration was tested. The duration of the accumulation
phase is 4 days.
}
\references{
Ciccia, T. (2019). Accumulation et devenir du mercure chez
l'espèce sentinelle Gammarus fossarum : de l'expérimentation au développement
d'un modèle toxicocinétique multi-compartiments.
Rapport de stage de Master 2, INRAE.
}
\keyword{dataset}
