% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_traces_fit.R
\name{mcmcTraces}
\alias{mcmcTraces}
\title{Traces of MCMC iterations}
\usage{
mcmcTraces(fit, plots = "all")
}
\arguments{
\item{fit}{An object of class \code{fitTK}}

\item{plots}{A string selecting the parameters. Defaults is \code{"all"} and select all parameters.
Deterministc parameters can be selected by setting \code{"deterministic"} and
stochastic parameter with \code{"stochastic"}}
}
\value{
A traceplot of class \code{ggplot}.
}
\description{
Traces of MCMC iterations
}
