% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data.R
\name{modelData}
\alias{modelData}
\alias{modelData.data.frame}
\title{Create a list giving data and parameters to use in the model inference.}
\usage{
modelData(object, ...)

\method{modelData}{data.frame}(object, time_accumulation, ...)
}
\arguments{
\item{object}{An object of class \code{data.frame}}

\item{\dots}{Further arguments to be passed to generic methods}

\item{time_accumulation}{A scalar givin accumulation time}
}
\value{
A \code{list} with data and parameters require for model inference.
}
\description{
Create a list giving data and parameters to use in the model inference.
}
