% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data_ode.R
\name{modelData_ode}
\alias{modelData_ode}
\title{Create a list giving data and parameters to use in the model inference.}
\usage{
modelData_ode(
  df_exposure,
  df_internal,
  y0 = 1,
  t0 = -0.001,
  unifMax = 10,
  time_accumulation = NULL,
  minK = -5,
  maxK = 5,
  ...
)

modelData_ode(
  df_exposure,
  df_internal,
  y0 = 1,
  t0 = -0.001,
  unifMax = 10,
  time_accumulation = NULL,
  minK = -5,
  maxK = 5,
  ...
)
}
\arguments{
\item{df_exposure}{Dataframe of exposure with 2 column (\code{time} and \code{value})}

\item{df_internal}{Dataframe of internal concentration with 2 column (\code{time} and \code{value})}

\item{y0}{Initial concentration}

\item{t0}{initial time point}

\item{unifMax}{Hyperparameter value}

\item{time_accumulation}{Time of accumulation}

\item{minK}{Hyperparameter value}

\item{maxK}{Hyperparameter value}

\item{\dots}{Additional arguments}
}
\value{
A \code{list} with data and parameters require for model inference.
}
\description{
Create a list giving data and parameters to use in the model inference.
}
