% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioacc_metric.R
\name{plot.bioaccMetric}
\alias{plot.bioaccMetric}
\title{Plot function for object of class \code{bioaccMetric}}
\usage{
\method{plot}{bioaccMetric}(x, ...)
}
\arguments{
\item{x}{a data frame}

\item{\dots}{Additional arguments}
}
\value{
A plot of class \code{ggplot}
}
\description{
Plot function for object of class \code{bioaccMetric}
}
