% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fit.R
\name{plot.fitTK}
\alias{plot.fitTK}
\title{Plotting method for \code{fitTK} objects}
\usage{
\method{plot}{fitTK}(x, time_interp = NULL, ...)
}
\arguments{
\item{x}{And object returned by fitTK}

\item{time_interp}{A vector with additional time point to interpolate.
Time point of the original data set are conserved.}

\item{\dots}{Additional arguments}
}
\value{
a plot of class \code{ggplot}
}
\description{
This is the generic \code{plot} S3 method for the
\code{fitTK}.  It plots the fit obtained for each
variable in the original dataset.
}
