% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exposure.R
\name{plot_exposure}
\alias{plot_exposure}
\title{Plot exposure profile}
\usage{
plot_exposure(object)
}
\arguments{
\item{object}{a data frame with exposure column}
}
\value{
a plot of class \code{ggplot}
}
\description{
Plot exposure profile
}
