% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PP.R
\name{plot_PriorPost}
\alias{plot_PriorPost}
\alias{plot_PriorPost.fitTK}
\title{Plot Posterior over Prior}
\usage{
plot_PriorPost(fit, ...)

\method{plot_PriorPost}{fitTK}(fit, select = "all", ...)
}
\arguments{
\item{fit}{An object of class \code{fitTK} returned by the function \code{fitTK()}.}

\item{\dots}{addition arguments}

\item{select}{A string selecting the parameters. Defaults is \code{"all"} and select all parameters.
Deterministic parameters can be selected by setting \code{"deterministic"} and
stochastic parameter with \code{"stochastic"}.}
}
\value{
An object of class \code{data.frame}.
}
\description{
Plot Posterior over Prior

Plot Posterior over Prior
}
