% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_fit.R
\name{predict.fitTK}
\alias{predict.fitTK}
\title{Prediction function}
\usage{
\method{predict}{fitTK}(fit, data)
}
\arguments{
\item{fit}{An object of \code{stanfit}}

\item{data}{A data set with one column \code{time} and 1 to 4 exposure
columns with name in \code{expw}, \code{exps}, \code{expf} and \code{exppw}}
}
\description{
Prediction function
}
