% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_fit.R
\name{quantile_table}
\alias{quantile_table}
\title{Quantiles of parameters}
\usage{
quantile_table(fit, probs = c(0.025, 0.5, 0.975))
}
\arguments{
\item{fit}{An object of class \code{fitTK}}

\item{probs}{Scalar or Vector of quantiles.
Default is 0.025, 0.5 and 0.975 giving median and 95\% credible interval}
}
\value{
A data frame with quantiles
}
\description{
Quantiles of parameters
}
