% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioacc-package.R
\docType{package}
\name{rbioacc-package}
\alias{rbioacc-package}
\alias{rbioacc}
\title{The 'rbioacc' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version NA. https://mc-stan.org
}
