% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_internals.R
\name{replace_}
\alias{replace_}
\title{Replace element of a vector}
\usage{
replace_(x, from, to)
}
\arguments{
\item{x}{a vector}

\item{from}{a vector of elements to replace}

\item{to}{a vector with replacing elements}
}
\value{
a vector
}
\description{
Replace element of a vector
}
\examples{
 
replace_(1:10,c(2,4,5,8), c(0,0,0,0))
replace_(c(1,2,2,3,2),c(3,2), c(4,5))


}
