% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t95.R
\name{t95}
\alias{t95}
\title{Return the time at 95\% depuration of the parent component}
\usage{
t95(fit)
}
\arguments{
\item{fit}{An object of class \code{fitTK}}
}
\value{
a numeric object
}
\description{
Return the time at 95\% depuration of the parent component
}
