% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic_fit.R
\name{waic}
\alias{waic}
\title{Widely Applicable Information Criterion (WAIC)}
\usage{
waic(fit)
}
\arguments{
\item{fit}{An object of class \code{fitTK}}
}
\value{
A numeric containing the WAIC
}
\description{
Compute WAIC using the \code{waic()} method of the loo package.
}
