% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioacc_metric.R
\name{bioacc_metric}
\alias{bioacc_metric}
\alias{bioacc_metric.fitTK}
\title{Biaccumulation metrics}
\usage{
bioacc_metric(fit, ...)

\method{bioacc_metric}{fitTK}(fit, type = "k", route = "all", ...)
}
\arguments{
\item{fit}{An \code{stanFit} object}

\item{\dots}{Further arguments to be passed to generic methods}

\item{type}{A string with the type of metric: \code{k} for the kinetics
BioConcentration Factor, \code{ss} for the steady state BioConcentration Factor.}

\item{route}{Provide exposure route: \code{all}}
}
\value{
a data frame
}
\description{
Biaccumulation metrics
}
