% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data.R
\name{.index_col_exposure}
\alias{.index_col_exposure}
\title{Return column matching "expw", "exps", "expf", "exppw" of a \code{data.frame}}
\usage{
.index_col_exposure(data_frame)
}
\arguments{
\item{x}{A vector}

\item{y}{A vector}
}
\description{
Return column matching "expw", "exps", "expf", "exppw" of a \code{data.frame}
}
