% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTK.R
\name{fitTK}
\alias{fitTK}
\alias{fitTK.stanTKdataCST}
\alias{fitTK.stanTKdataVAR}
\title{Posterior predictive check}
\usage{
fitTK(stanTKdata, ...)

\method{fitTK}{stanTKdataCST}(stanTKdata, ...)

\method{fitTK}{stanTKdataVAR}(stanTKdata, ...)
}
\arguments{
\item{stanTKdata}{List of Data require for computing}

\item{\dots}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{fitTK} containing two object: \code{stanTKdata}
the data set used for inference and \code{stanfit}  returned by \code{rstan::sampling}
}
\description{
Posterior predictive check

Bayesian inference of TK model with Stan

Bayesian inference of TK model with variable exposure profile (BETA version)
}
