% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fit.R
\name{plot.fitTK}
\alias{plot.fitTK}
\title{Plotting method for \code{fitTK} objects}
\usage{
\method{plot}{fitTK}(fit)
}
\arguments{
\item{fit}{And object returned by fitTK}
}
\description{
This is the generic \code{plot} S3 method for the
\code{fitTK}.  It plots the fit obtained for each
variable in the original dataset.
}
