% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_predict.R
\name{plot.predictTK}
\alias{plot.predictTK}
\title{Plotting method for \code{predictTK} objects}
\usage{
\method{plot}{predictTK}(x, ...)
}
\arguments{
\item{x}{An object of class \code{predictTK} returned by predict}

\item{\dots}{Additional arguments}
}
\description{
This is the generic \code{plot} S3 method for the
\code{predictTK}.
}
