% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PP.R
\name{plot_PriorPost}
\alias{plot_PriorPost}
\alias{plot_PriorPost.fitTK}
\alias{plot_PriorPost.df_PP}
\title{Plot Posterior over Prior}
\usage{
plot_PriorPost(x, ...)

\method{plot_PriorPost}{fitTK}(x, select = "all", ...)

\method{plot_PriorPost}{df_PP}(x, select = "all", ...)
}
\arguments{
\item{x}{A data.frame of class \code{df_PP} returned by the function \code{df_PriorPost()}.}

\item{\dots}{addition arguments}

\item{select}{A string selecting the parameters. Defaults is \code{"all"} and select all parameters.
Deterministic parameters can be selected by setting \code{"deterministic"} and
stochastic parameter with \code{"stochastic"}.}
}
\value{
A plot of class \code{ggplot}.

A plot of class \code{ggplot}.
}
\description{
Plot Posterior over Prior

Plot Posterior over Prior
}
