% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrf_fit.R
\name{psrf}
\alias{psrf}
\title{Potential Scale Reduction Factors (PSRF) of the parameters}
\usage{
psrf(fit)
}
\arguments{
\item{fit}{An object of class \code{fitTK}}
}
\value{
An object of class \code{data.frame} with two columns: PSRF and parameter

a data frame with Potential Scale Reduction Factors
}
\description{
Potential Scale Reduction Factors (PSRF) of the parameters
}
