% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioacc-package.R
\docType{package}
\name{rbioacc-package}
\alias{rbioacc-package}
\alias{rbioacc}
\title{The 'rbioacc' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version NA. https://mc-stan.org
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.in2p3.fr/mosaic-software/rbioacc}
  \item Report bugs at \url{https://gitlab.in2p3.fr/mosaic-software/rbioacc/-/issues}
}

}
\author{
\strong{Maintainer}: Virgile Baudrot \email{virgile.baudrot@qonfluens.com}

Authors:
\itemize{
  \item Sandrine Charles
  \item Miléna Kaag
  \item Aude Ratier
  \item Aurélie Siberchicot
}

Other contributors:
\itemize{
  \item Ophélia Gestin [contributor]
  \item Christelle Lopes [contributor]
  \item Gauthier Multari [contributor]
  \item Alain Pavé [contributor]
}

}
